Rename class CommandFormat
