+++ src/main/java/seedu/duke/logic/parser/DeleteCommandParser.java
@@ -8,10 +8,10 @@ import seedu.duke.logic.commands.module.DeleteModuleCommand;
 import seedu.duke.logic.commands.task.DeleteTaskCommand;
 import seedu.duke.logic.parser.exceptions.ParseException;
 
-import static seedu.duke.commons.core.CommandFormats.DELETE_TASK_FORMAT;
-import static seedu.duke.commons.core.CommandFormats.DELETE_LESSON_FORMAT;
-import static seedu.duke.commons.core.CommandFormats.DELETE_MODULE_FORMAT;
-import static seedu.duke.commons.core.CommandFormats.promptFormat;
+import static seedu.duke.commons.core.CommandFormat.DELETE_TASK_FORMAT;
+import static seedu.duke.commons.core.CommandFormat.DELETE_LESSON_FORMAT;
+import static seedu.duke.commons.core.CommandFormat.DELETE_MODULE_FORMAT;
+import static seedu.duke.commons.core.CommandFormat.promptFormat;
 import static seedu.duke.logic.parser.ParserUtil.parseCommandType;
 import static seedu.duke.logic.parser.ParserUtil.parseToZeroIndex;
 import static seedu.duke.logic.parser.ParserUtil.removeFirstParam;
