+++ src/main/java/seedu/duke/logic/parser/ListCommandParser.java
@@ -2,17 +2,16 @@ package seedu.duke.logic.parser;
 
 import seedu.duke.commons.core.CommandType;
 import seedu.duke.commons.core.DayOfTheWeek;
-import seedu.duke.commons.core.Messages;
 import seedu.duke.logic.commands.Command;
 import seedu.duke.logic.commands.lesson.ListLessonCommand;
 import seedu.duke.logic.commands.module.ListModuleCommand;
 import seedu.duke.logic.commands.task.ListTaskCommand;
 import seedu.duke.logic.parser.exceptions.ParseException;
 
-import static seedu.duke.commons.core.CommandFormats.LIST_TASK_FORMAT;
-import static seedu.duke.commons.core.CommandFormats.LIST_LESSON_FORMAT;
-import static seedu.duke.commons.core.CommandFormats.LIST_MODULE_FORMAT;
-import static seedu.duke.commons.core.CommandFormats.promptFormat;
+import static seedu.duke.commons.core.CommandFormat.LIST_TASK_FORMAT;
+import static seedu.duke.commons.core.CommandFormat.LIST_LESSON_FORMAT;
+import static seedu.duke.commons.core.CommandFormat.LIST_MODULE_FORMAT;
+import static seedu.duke.commons.core.CommandFormat.promptFormat;
 import static seedu.duke.commons.util.DayUtil.isToday;
 import static seedu.duke.commons.util.DayUtil.isTomorrow;
 import static seedu.duke.logic.parser.ParserUtil.isVerbose;
